tero-gae/requirements/dev.txt
nose
ipython
django-extensions
Werkzeug
